+++ api/src/main/java/io/gatehill/imposter/config/ResolvedResourceConfig.java
@@ -11,11 +11,13 @@ public class ResolvedResourceConfig {
     private final RestResourceConfig config;
     private final Map<String, String> pathParams;
     private final Map<String, String> queryParams;
+    private final Map<String, String> requestHeaders;
 
-    public ResolvedResourceConfig(RestResourceConfig config, Map<String, String> pathParams, Map<String, String> queryParams) {
+    public ResolvedResourceConfig(RestResourceConfig config, Map<String, String> pathParams, Map<String, String> queryParams, Map<String, String> requestHeaders) {
         this.config = config;
         this.pathParams = pathParams;
         this.queryParams = queryParams;
+        this.requestHeaders = requestHeaders;
     }
 
     public RestResourceConfig getConfig() {
@@ -28,5 +30,9 @@ public class ResolvedResourceConfig {
 
     public Map<String, String> getQueryParams() {
         return queryParams;
+    }
+
+    public Map<String, String> getRequestHeaders() {
+        return requestHeaders;
     }
 }