+++ api/src/main/java/io/gatehill/imposter/script/ExecutionContext.java
@@ -1,8 +1,9 @@
 package io.gatehill.imposter.script;
 
+import io.gatehill.imposter.util.StringUtil;
+
 import java.util.Map;
 import java.util.function.Supplier;
-import java.util.stream.Collectors;
 
 import static java.util.Objects.isNull;
 import static java.util.Optional.ofNullable;
@@ -109,9 +110,7 @@ public class ExecutionContext {
          * @return the {@link #getHeaders()} map, but with all keys in lowercase
          */
         public Map<String, String> getNormalisedHeaders() {
-            return getHeaders().entrySet().stream().collect(
-                    Collectors.toMap(e -> e.getKey().toLowerCase(), Map.Entry::getValue)
-            );
+            return StringUtil.convertKeysToLowerCase(getHeaders());
         }
 
         /**